Jeff's Changes + Sticky Footer + Others
Jeff changed a few things so the spacing is more consistent and
professional looking.
I added some CSS to make it so the footer is a constant height, it also
forces the footer to the ground.
Additionally, the paypal button has been implemented. We're using the
default right now, but honestly it looks pretty good.
Finally, carousel "active buttons" work properly now.